#86 refactor : BottomSheet 함수 -> class 변경
